Allow EVs to be set while calling examples
Currently, in order to use the provided example scripts, any environment
variables listed at the top of the files need to be manually edited and
saved before the script can be run. This limits the ability to automate
and streamline the process of running these scripts and can be tedious to
create copies of files while running from a shared filesystem, such as an
NFS, on multiple nodes as the values will be overwritten on different
nodes.

By using default values and allowing the variables to be read into the
script, it is much easier to call scripts with different parameters on
multiple nodes, such as:

```
MASTER_ADDR=127.0.0.1 examples/pretrain_bert_distributed.sh
```

Signed-Off-By: Robert Clark <roclark@nvidia.com>