Support S3 data loading
This commit introduces the S3IndexedDataset, which supports
loading a dataset stored in S3 in the same format as the
MMapIndexedDataset. In particular, the .idx file is downloaded
to a local directory at initialization so that we can memory map
it and the .bin file is streamed into memory block-by-block.