Support cloud storage for the torch strategy in dist checkpointing
This commit adds support for saving checkpoints to cloud storage
(e.g., S3) and loading checkpoints from cloud storage for the
torch strategy in distributed checkpointing. It does so by
replacing pathlib.Path with cloudpathlib.AnyPath, FileSystemReader
with FsspecSystemReader, and FileSytemWriter with
FsspecSystemWriter.

The commit enables cloud checkpointing, but makes little attempt
to optimize it.