Fix BlendableDataset for low sampling probs
When supplying low enough sampling probabilities for some datasets so
that they won't be included in the blend, this part could incorrectly
raise an error because `sizes` was potentially compared with an
incorrect dataset which was never part of the blend to begin with.

We now make sure that the dataset is actually part of the blend by only
iterating over those indices that are present at all.