ci: Fix dev branch CI (#1922)
Fix dev branch CI 

For some reason, on the dev branch, the call to `energy_monitor.pause()`
fails in the training script. It does not seem to be related to the
dependencies because this still fails when using the same docker image
with same pyproject.toml and uv.lock file. I recommend we merge this to
unblock the dev branch and allow us more time to dig deeper into the
root cause.

---------

Signed-off-by: Charlie Truong <chtruong@nvidia.com>
Co-authored-by: Oliver Koenig <okoenig@nvidia.com>